@@ -9,7 +9,10 @@ import pl.mosenko.sunnypodlaskie.util.WeatherDataJobSyncUtils
  * Created by syk on 26.05.17.
  */
 class AlarmWeatherSyncReceiver : BroadcastReceiver() {
-    override fun onReceive(context: Context?, intent: Intent?) {
-        WeatherDataJobSyncUtils.scheduleFirebaseJobDispatcherSync(context)
+
+    override fun onReceive(context: Context, intent: Intent) {
+        if (Intent.ACTION_BOOT_COMPLETED == intent.action) {
+            WeatherDataJobSyncUtils.scheduleFirebaseJobDispatcherSync(context)
+        }
     }
 }
