@@ -9,9 +9,9 @@ import pl.mosenko.sunnypodlaskie.util.WeatherDataAlarmSyncUtil
  * Created by syk on 26.05.17.
  */
 class DeviceBroadcastReceiver : BroadcastReceiver() {
-    override fun onReceive(context: Context?, intent: Intent?) {
-        val action = intent.getAction()
-        if (Intent.ACTION_BOOT_COMPLETED == action) {
+
+    override fun onReceive(context: Context, intent: Intent) {
+        if (Intent.ACTION_BOOT_COMPLETED == intent.action) {
             WeatherDataAlarmSyncUtil.startAlarm(context)
         }
     }
