@@ -1,13 +1,13 @@
 package pl.mosenko.sunnypodlaskie.mvp.weatherdatalist
 
 import android.content.Context
-import android.support.v7.widget.RecyclerView
-import android.support.v7.widget.RecyclerView.ViewHolder
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
+import androidx.recyclerview.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView.ViewHolder
 import butterknife.BindView
 import butterknife.ButterKnife
 import pl.mosenko.sunnypodlaskie.R
@@ -19,67 +19,60 @@ import java.util.*
 /**
  * Created by syk on 15.05.17.
  */
-class WeatherDataListAdaper(private val context: Context?, clickHandler: WeatherDataClickedListener?) : RecyclerView.Adapter<WeatherViewHolder?>() {
-    private var weatherDataList: MutableList<WeatherDataEntity?>?
-    private val clickHandler: WeatherDataClickedListener?
-    override fun onCreateViewHolder(parent: ViewGroup?, viewType: Int): WeatherViewHolder? {
+class WeatherDataListAdaper(private val context: Context?, val clickHandler: WeatherDataClickedListener) : RecyclerView.Adapter<WeatherViewHolder>() {
+
+    private var weatherDataList: MutableList<WeatherDataEntity> = ArrayList()
+
+    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): WeatherViewHolder {
         val itemView = LayoutInflater.from(context).inflate(R.layout.weather_data_list_item, parent, false)
         itemView.isFocusable = true
         return WeatherViewHolder(itemView)
     }
 
-    override fun onBindViewHolder(holder: WeatherViewHolder?, position: Int) {
+    override fun onBindViewHolder(holder: WeatherViewHolder, position: Int) {
         val weatherInfo = weatherDataList.get(position)
         val iconResource = WeatherDataUtil.getWeatherIconResourceByCode(weatherInfo.getIconKey())
         holder.weatherIcon.setImageResource(iconResource)
         holder.city.setText(weatherInfo.getCity())
         holder.temperature.setText(WeatherDataUtil.getFormattedTemperature(weatherInfo.getTemperature()))
-        holder.weatherDescription.setText(weatherInfo.getWeatherCondition().description)
+        holder.weatherDescription.setText(weatherInfo.getWeatherCondition().getDescription())
     }
 
     override fun getItemCount(): Int {
         return weatherDataList.size
     }
 
-    fun swapWeatherList(weatherList: MutableList<WeatherDataEntity?>?) {
+    fun swapWeatherList(weatherList: MutableList<WeatherDataEntity>) {
         weatherDataList = weatherList
         notifyDataSetChanged()
     }
 
     interface WeatherDataClickedListener {
-        open fun onWeatherDataItemClick(id: Long)
+        fun onWeatherDataItemClick(id: Long)
     }
 
-    inner class WeatherViewHolder(itemView: View?) : ViewHolder(itemView), View.OnClickListener {
-        @kotlin.jvm.JvmField
+    inner class WeatherViewHolder(itemView: View) : ViewHolder(itemView), View.OnClickListener {
         @BindView(R.id.WeatherDataListFragment_icon_weather)
-        var weatherIcon: ImageView? = null
+        lateinit var weatherIcon: ImageView
 
-        @kotlin.jvm.JvmField
         @BindView(R.id.WeatherDataListFragment_city)
-        var city: TextView? = null
+        lateinit var city: TextView
 
-        @kotlin.jvm.JvmField
         @BindView(R.id.WeatherDataListFragment_weather_description)
-        var weatherDescription: TextView? = null
+        lateinit var weatherDescription: TextView
 
-        @kotlin.jvm.JvmField
         @BindView(R.id.WeatherDataListFragment_temperature)
-        var temperature: TextView? = null
-        override fun onClick(v: View?) {
+        lateinit var temperature: TextView
+
+        override fun onClick(v: View) {
             val adapterPosition = adapterPosition
             val weatherDataEntity = weatherDataList.get(adapterPosition)
-            clickHandler.onWeatherDataItemClick(weatherDataEntity.getId())
+            clickHandler.onWeatherDataItemClick(weatherDataEntity.getId()!!)
         }
 
         init {
             ButterKnife.bind(this, itemView)
             itemView.setOnClickListener(this)
         }
-    }
-
-    init {
-        weatherDataList = ArrayList()
-        this.clickHandler = clickHandler
     }
 }
