@@ -10,5 +10,5 @@ import retrofit2.http.Query
  */
 interface WeatherDataAPI {
     @GET("data/2.5/group?id=757718,759503,776597,771506,768785,766027,764421,758651,753895,754479,775986,759320,771158,776069&units=metric")
-    open fun getCurrentWeatherData(@Query("appid") apiKey: String?): Observable<WeatherDataDto?>?
+    fun getCurrentWeatherData(@Query("appid") apiKey: String): Observable<WeatherDataDto>
 }
