@@ -4,5 +4,5 @@ package pl.mosenko.sunnypodlaskie.persistence.entities
  * Created by syk on 06.06.17.
  */
 interface BaseOrmLiteEntity {
-    open fun getId(): Long?
+    fun getId(): Long?
 }
