@@ -4,7 +4,6 @@ import android.provider.BaseColumns
 import com.j256.ormlite.field.DatabaseField
 import com.j256.ormlite.table.DatabaseTable
 import pl.mosenko.sunnypodlaskie.persistence.dao.WeatherDataEntityDAO
-import pl.mosenko.sunnypodlaskie.persistence.entities.WeatherDataEntity
 import java.util.*
 
 /**
@@ -44,35 +43,35 @@ class WeatherDataEntity : AbstractOrmLiteEntity() {
 
     @DatabaseField(columnName = SUNSET_COLUMN)
     private var sunset: Date? = null
-    fun getCity(): String? {
-        return city
+    fun getCity(): String {
+        return city!!
     }
 
-    fun setCity(city: String?) {
+    fun setCity(city: String) {
         this.city = city
     }
 
-    fun getReceivingTime(): Date? {
-        return receivingTime
+    fun getReceivingTime(): Date {
+        return receivingTime!!
     }
 
-    fun setReceivingTime(receivingTime: Date?) {
+    fun setReceivingTime(receivingTime: Date) {
         this.receivingTime = receivingTime
     }
 
-    fun getIconKey(): String? {
-        return iconKey
+    fun getIconKey(): String {
+        return iconKey!!
     }
 
-    fun setIconKey(iconKey: String?) {
+    fun setIconKey(iconKey: String) {
         this.iconKey = iconKey
     }
 
-    fun getWeatherCondition(): WeatherConditionEntity? {
-        return weatherCondition
+    fun getWeatherCondition(): WeatherConditionEntity {
+        return weatherCondition!!
     }
 
-    fun setWeatherCondition(weatherCondition: WeatherConditionEntity?) {
+    fun setWeatherCondition(weatherCondition: WeatherConditionEntity) {
         this.weatherCondition = weatherCondition
     }
 
@@ -149,17 +148,17 @@ class WeatherDataEntity : AbstractOrmLiteEntity() {
     }
 
     companion object {
-        val TABLE_NAME: String? = "weather_data"
-        val CITY_COLUMN: String? = "city"
-        val RECEIVING_TIME_COLUMN: String? = "receiving_time"
-        val ICON_KEY_COLUMN: String? = "icon_key"
-        val WEATHER_CONDITION: String? = "weather_condition"
-        val TEMPERATURE_COLUMN: String? = "temperature"
-        val PRESSURE_COLUMN: String? = "pressure"
-        val HUMIDITY_COLUMN: String? = "humidity"
-        val WIND_SPEED_COLUMN: String? = "wind_speed"
-        val WIND_DEGREE_COLUMN: String? = "wind_degree"
-        val SUNRISE_COLUMN: String? = "sunrise"
-        val SUNSET_COLUMN: String? = "sunset"
+        const val TABLE_NAME: String = "weather_data"
+        const val CITY_COLUMN: String = "city"
+        const val RECEIVING_TIME_COLUMN: String = "receiving_time"
+        const val ICON_KEY_COLUMN: String = "icon_key"
+        const val WEATHER_CONDITION: String = "weather_condition"
+        const val TEMPERATURE_COLUMN: String = "temperature"
+        const val PRESSURE_COLUMN: String = "pressure"
+        const val HUMIDITY_COLUMN: String = "humidity"
+        const val WIND_SPEED_COLUMN: String = "wind_speed"
+        const val WIND_DEGREE_COLUMN: String = "wind_degree"
+        const val SUNRISE_COLUMN: String = "sunrise"
+        const val SUNSET_COLUMN: String = "sunset"
     }
 }
