@@ -11,38 +11,37 @@ import pl.mosenko.sunnypodlaskie.persistence.dao.WeatherDataEntityDAO
 import pl.mosenko.sunnypodlaskie.persistence.entities.WeatherDataEntity
 import pl.mosenko.sunnypodlaskie.util.WeatherDtoEntityConverter
 import java.sql.SQLException
-import javax.inject.Inject
 
 /**
  * Created by syk on 03.06.17.
  */
-class WeatherDataRepository(@field:Inject var rxWeatherDataAPI: RxWeatherDataAPI?, @field:Inject var weatherDataEntityDAO: WeatherDataEntityDAO?, @field:Inject var weatherConditionEntities: WeatherConditionEntityDAO?) {
-    fun loadCurrentWeatherData(isConnectedToInternet: Boolean, callback: Callback?): Disposable? {
-        val weatherDataEntityObservable: Observable<MutableList<WeatherDataEntity?>?>?
-        weatherDataEntityObservable = if (isConnectedToInternet) {
+class WeatherDataRepository(var rxWeatherDataAPI: RxWeatherDataAPI, var weatherDataEntityDAO: WeatherDataEntityDAO, var weatherConditionEntities: WeatherConditionEntityDAO) {
+
+    fun loadCurrentWeatherData(isConnectedToInternet: Boolean, callback: Callback): Disposable {
+        val weatherDataEntityObservable: Observable<MutableList<WeatherDataEntity>> = if (isConnectedToInternet) {
             rxWeatherDataAPI.getCurrentWeatherData()
-                    .map { obj: WeatherDataDto? -> WeatherDtoEntityConverter.convertToWeatherDataEntityList() }
-                    .doOnNext { weatherDataEntityList: MutableList<WeatherDataEntity?>? -> cacheCurrentWeatherData(weatherDataEntityList) }
+                    .map { obj: WeatherDataDto -> WeatherDtoEntityConverter.convertToWeatherDataEntityList(obj) }
+                    .doOnNext { weatherDataEntityList: MutableList<WeatherDataEntity> -> cacheCurrentWeatherData(weatherDataEntityList) }
         } else {
             weatherDataEntityDAO.rxQueryForAll()
         }
         return weatherDataEntityObservable
                 .subscribeOn(Schedulers.io())
                 .observeOn(AndroidSchedulers.mainThread())
-                .subscribe({ p: MutableList<WeatherDataEntity?>? -> callback.onNextWeatherDataEntities(p, isConnectedToInternet) }) { throwable: Throwable? -> callback.onError(throwable) }
+                .subscribe({ p: MutableList<WeatherDataEntity> -> callback.onNextWeatherDataEntities(p, isConnectedToInternet) }) { throwable: Throwable -> callback.onError(throwable) }
     }
 
     @Throws(SQLException::class)
-    private fun cacheCurrentWeatherData(weatherDataEntityList: MutableList<WeatherDataEntity?>?) {
+    private fun cacheCurrentWeatherData(weatherDataEntityList: MutableList<WeatherDataEntity>) {
         weatherDataEntityDAO.deleteBuilder().delete()
-        weatherDataEntityDAO.create(weatherDataEntityList)
+        weatherDataEntityDAO.create(weatherDataEntityList.toList())
         for (weatherEntity in weatherDataEntityList) {
             weatherConditionEntities.refresh(weatherEntity.getWeatherCondition())
         }
     }
 
     interface Callback {
-        open fun onNextWeatherDataEntities(weatherDataEntityList: MutableList<WeatherDataEntity?>?, isConnectedToInternet: Boolean)
-        open fun onError(throwable: Throwable?)
+        fun onNextWeatherDataEntities(weatherDataEntityList: MutableList<WeatherDataEntity>, isConnectedToInternet: Boolean)
+        fun onError(throwable: Throwable)
     }
 }
