@@ -1,43 +1,44 @@
 package pl.mosenko.sunnypodlaskie.util
 
 import android.content.Context
-import android.support.v7.preference.PreferenceManager
 import android.util.Log
+import androidx.preference.PreferenceManager
 import pl.mosenko.sunnypodlaskie.BuildConfig
 import pl.mosenko.sunnypodlaskie.R
 import java.text.ParseException
 import java.text.SimpleDateFormat
 import java.util.*
+import kotlin.jvm.Throws
 
 /**
  * Created by syk on 26.05.17.
  */
 object PreferenceWeatherUtil {
     private val TAG = PreferenceWeatherUtil::class.java.simpleName
+
     fun areNotificationsEnabled(context: Context): Boolean {
         val displayNotificationsKey = context.getString(R.string.pref_enable_notifications_key)
-        val showNotificationsDefaultValue = context.resources.getBoolean(R.bool.show_notifications_by_default)
+        val showNotificationsDefaultValue =
+            context.resources.getBoolean(R.bool.show_notifications_by_default)
         val sharedPreferences = PreferenceManager.getDefaultSharedPreferences(context)
         return sharedPreferences.getBoolean(displayNotificationsKey, showNotificationsDefaultValue)
     }
 
-    fun getSyncDate(context: Context): Date? {
+    fun getSyncDate(context: Context): Date {
         val syncTimeString = getStringSyncTime(context)
         return parseSyncTime(syncTimeString)
     }
 
-    private fun getStringSyncTime(context: Context?): String {
+    private fun getStringSyncTime(context: Context): String {
         val sharedPreferences = PreferenceManager.getDefaultSharedPreferences(context)
-        val syncTimeString: String?
-        syncTimeString = if (sharedPreferences.contains(context.getString(R.string.pref_sync_time_key))) {
-            sharedPreferences.getString(context.getString(R.string.pref_sync_time_key), "")
+        return if (sharedPreferences.contains(context.getString(R.string.pref_sync_time_key))) {
+            sharedPreferences.getString(context.getString(R.string.pref_sync_time_key), "")!!
         } else {
             context.getString(R.string.pref_sync_time_default)
         }
-        return syncTimeString
     }
 
-    private fun parseSyncTime(syncTimeString: String?): Date? {
+    private fun parseSyncTime(syncTimeString: String): Date {
         val syncTimeDate = parseSyncTimeToPartiallyDate(syncTimeString)
         val syncDateCalendar = getCurrentSyncTimeCalendar(syncTimeDate)
         return syncDateCalendar.time
@@ -45,7 +46,7 @@ object PreferenceWeatherUtil {
 
     private fun getCurrentSyncTimeCalendar(syncTimeDate: Date?): Calendar {
         val syncTimeCalendar = Calendar.getInstance()
-        syncTimeCalendar.time = syncTimeDate
+        syncTimeCalendar.time = syncTimeDate!!
         val syncDateCalendar = Calendar.getInstance()
         syncDateCalendar[Calendar.HOUR_OF_DAY] = syncTimeCalendar[Calendar.HOUR_OF_DAY]
         syncDateCalendar[Calendar.MINUTE] = syncTimeCalendar[Calendar.MINUTE]
@@ -54,7 +55,7 @@ object PreferenceWeatherUtil {
         return syncDateCalendar
     }
 
-    fun parseSyncTimeToPartiallyDate(syncTimeString: String?): Date? {
+    fun parseSyncTimeToPartiallyDate(syncTimeString: String): Date? {
         val simpleTimeFormat = SimpleDateFormat("HH:mm", Locale.getDefault())
         try {
             val syncTimeDate = simpleTimeFormat.parse(syncTimeString)
@@ -67,7 +68,7 @@ object PreferenceWeatherUtil {
         return null
     }
 
-    private fun isValidTime(syncTimeString: String?): Boolean {
+    private fun isValidTime(syncTimeString: String): Boolean {
         val hourPart = syncTimeString.substring(0, syncTimeString.indexOf(":"))
         val minutePart = syncTimeString.substring(syncTimeString.indexOf(":") + 1)
         val hours: Int?
@@ -78,6 +79,6 @@ object PreferenceWeatherUtil {
         } catch (ne: NumberFormatException) {
             return false
         }
-        return hours >= 0 && hours <= 24 && minutes >= 0 && minutes <= 59
+        return hours in 0..24 && minutes >= 0 && minutes <= 59
     }
 }
